Set vision for application and move to html
This project needs to move towards being a gem. The README has been
documented with the expected API for using this. It also takes the
step towards using html by default.

When testing, due to the asynchronous nature of background jobs,
actioncable and feature testing thing, both the background jobs and
actioncable management need to be handled off process. To keep
confusion low I kept with the delayed_job setup and use redis pubsub
on actioncable. This is more dependencies than I'd like, but it seems
required when testing out of process javascript tests.

* Update README with expected API
* Simplify rendering to be serverside and more similar to the `render` syntax
* For the time being do not implement the text interface